payment/requirements.txt
aioredis==2.0.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
cleo==1.0.0a4
click==8.1.2
crashtest==0.3.1
Deprecated==1.2.13
fastapi==0.75.1
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
packaging==21.3
pptree==3.1
pydantic==1.9.0
pylev==1.4.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.2.1
redis-om==0.0.21
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
types-redis==4.1.19
types-six==1.16.12
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
wrapt==1.14.0
